+++ test/integ/allocate/integ.allocate.ts.snapshot/tree.json
@@ -2901,7 +2901,7 @@
                                 {
                                   "Action": "execute-api:Invoke",
                                   "Condition": {
-                                    "ForAllValues:StringNotEquals": {
+                                    "ForAllValues:ArnNotEquals": {
                                       "aws:PrincipalArn": [
                                         {
                                           "Fn::Join": [
@@ -2928,8 +2928,7 @@
                                   "Principal": {
                                     "AWS": "*"
                                   },
-                                  "Resource": "*",
-                                  "Sid": "DenySameAccountAccess"
+                                  "Resource": "*"
                                 }
                               ],
                               "Version": "2012-10-17"
@@ -2979,7 +2978,7 @@
                               "aws:cdk:cloudformation:type": "AWS::ApiGateway::Stage",
                               "aws:cdk:cloudformation:props": {
                                 "deploymentId": {
-                                  "Ref": "AtmosphereEndpointApiDeployment2DF1AFBE25d4957745e9fa2d92c7db6fbe1a5d3a"
+                                  "Ref": "AtmosphereEndpointApiDeployment2DF1AFBE8a51507bf59673cf8318352f428c7177"
                                 },
                                 "restApiId": {
                                   "Ref": "AtmosphereEndpointApi64E91738"
@@ -3940,7 +3939,7 @@
                       "s3Bucket": {
                         "Fn::Sub": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}"
                       },
-                      "s3Key": "54818df61121ecd64e09be30b089f337bbc3146e3a5732243354a7e955b03762.zip"
+                      "s3Key": "32f9cbdd00c906811ccd5dca4625101aea15507dff070c9f62ba42260a07d382.zip"
                     },
                     "description": "test/integ/allocate/assert.lambda.ts",
                     "environment": {
